--- bulbControlFrontend/Realtime_PyAudio_FFT/requirements.txt
@@ -1,5 +1,0 @@
-numpy
-matplotlib
-scipy
-pygame
-PyAudio
