+++ bulbControlFrontend/static/js/generic.js
@@ -1,3 +1,10 @@
+window.onload = function () {
+    var error_wrapper = document.getElementById('error-wrapper');
+    error_wrapper.addEventListener('hidden.bs.modal', function () {
+        fetch('/clearError/');
+    });
+};
+
 async function audioSync(element) {
     var audio_device = document.getElementById('audio_device_selector').value;
     var button = document.getElementById('activate-music-sync-button');
@@ -11,7 +18,7 @@ async function audioSync(element) {
         button.classList.add('btn-outline-danger');
         request = '/stopSync/';
     }
-    response = await fetch('/activateSync/', {
+    response = await fetch(request, {
         method: 'post',
         body: audio_device,
     }).then(function (response) {
@@ -19,15 +26,28 @@ async function audioSync(element) {
     });
 
     if (response['result']) {
-        element.offsetParent.style.visibility = 'hidden';
     } else {
         var errorMessage = JSON.parse(
             document.getElementById('errorMessage').textContent,
         );
         errorMessage = 'Audio sync failed to start, please try again';
-        error_wrapper = document.getElementById('error-wrapper');
-        error_wrapper.style.visibility = 'visible';
-        error_message = error_wrapper.getElementsByTagName('span')[0];
+        var error_wrapper = new bootstrap.Modal(
+            document.getElementById('error-wrapper'),
+            {
+                keyboard: true,
+            },
+        );
+        error_wrapper.show();
+        error_message = document
+            .getElementById('error-wrapper')
+            .getElementsByTagName('p')[0];
         error_message.textContent = errorMessage;
     }
+}
+function toggleTheme() {
+    var html = document.documentElement;
+    html.setAttribute(
+        'data-bs-theme',
+        html.getAttribute('data-bs-theme') === 'dark' ? 'light' : 'dark',
+    );
 }
