+++ bulbControlFrontend/static/main.css
@@ -5,3 +5,17 @@ body {
 .fa-solid.fa-lightbulb {
     font-size: 10vw;
 }
+
+#edit-bulb-link:hover {
+    color: black;
+    font-weight: bolder;
+    background-color: lightblue;
+    transition: 0.5s;
+}
+
+#delete-bulb-link:hover {
+    color: black;
+    font-weight: bolder;
+    background-color: lightcoral;
+    transition: 0.5s;
+}
