+++ bulbControlFrontend/templates/bulb_crud.html
@@ -1,0 +1,36 @@
+{% extends "master.html" %} {% load static %} {% block title %}Bulb Bop{% endblock %} {% block content %}
+<div class="container-fluid">
+    <table class="table table-striped table-hover table-bordered">
+        <caption>Table of saved bulbs</caption>
+        <thead>
+            <tr>
+                <th scope="col">Bulb Name</th>
+                <th scope="col">Bulb Ip</th>
+                <th scope="col">Bulb State</th>
+                <th scope="col" colspan="2">Actions</th>
+            </tr>
+        </thead>
+        {% for x in bulbs %}
+        <tbody>
+            <tr>
+                <td>{{ x.bulbName }}</td>
+                <td>{{ x.bulbIp }}</td>
+                <td>{{ x.bulbState }}</td>
+                <td id="edit-bulb-link">Edit</td>
+                <td id="delete-bulb-link">
+                    <form action="/delete/{{x.bulbIp}}" method="post">
+                        <div class="container"><button type="submit" style="height: 100%; width: 100%; padding: 0; border: none; background: none">Delete</button></div>
+                    </form>
+                </td>
+            </tr>
+        </tbody>
+        {% endfor %}
+    </table>
+    <div class="row">
+        <div class="col-md-12">
+            <a href="/bulb/create">Create</a>
+        </div>
+    </div>
+</div>
+
+{% endblock %}
