--- bulbControlFrontend/templates/error.html
@@ -1,8 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <title>Wrong address</title>
-    <body>
-        <h1>Ooops!</h1>
-        <h2>I cannot find the file you requested!</h2>
-    </body>
-</html>
