+++ bulbControlFrontend/templates/generic_description.html
@@ -1,6 +1,6 @@
 <head>
     <meta name="author" content="Ryan McClean" />
-    <meta name="contact" content="https://github.com/RyanMcClean"
+    <meta name="contact" content="https://github.com/RyanMcClean" />
 </head>
 
 <p class="generic_description">
