+++ bulbControlFrontend/templates/master.html
@@ -18,7 +18,6 @@
         crossorigin="anonymous"
     ></script>
     <link rel="stylesheet" href="{% static 'main.css' %}" />
-    <script type="text/javascript" src="{% static 'js/errorHandler.js' %}"></script>
     <script type="text/javascript" src="{% static 'js/imageHandler.js' %}"></script>
     <script type="text/javascript" src="{% static 'js/generic.js' %}"></script>
     <head>
@@ -45,42 +44,85 @@
                         <span class="navbar-toggler-icon"></span>
                     </button>
                 </div>
-                    <div class="collapse container navbar-collapse" id="navbarSupportedContent">
-                        <div class="container-fluid justify-content-end text-end">
-                            <ul class="navbar-nav mx-auto mb-2 mb-lg-0">
-                                <div class="d-flex align-items-center">
-                                    <li class="nav-item">
-                                        <a id="home-link" class="nav-link active fs-4 fw-bold" aria-current="page" href="/">Home</a>
-                                    </li>
-                                    <li class="nav-item">
-                                        <a id="faq-link" class="nav-link disabled fs-4 fw-bold" aria-disabled="true" href="/">FAQs</a>
-                                    </li>
-                                    <li class="nav-item">
-                                        <a id="about-link" class="nav-link disabled fs-4 fw-bold" aria-disabled="true" href="/">About</a>
-                                    </li>
-                                    <li class="nav-item">
-                                        <a id="find-bulbs-button" class="btn btn-outline-primary fw-bolder" role="button" title="Find Bulbs" href="/discover">Find Bulbs</a>
-                                    </li>
-                                </div>
-                            </ul>
-                        </div>
+                <div class="collapse container navbar-collapse" id="navbarSupportedContent">
+                    <div class="container-fluid justify-content-end text-end">
+                        <ul class="navbar-nav mx-auto mb-2 mb-lg-0">
+                            <div class="d-flex align-items-center px-5">
+                                <li class="nav-item">
+                                    <a id="home-link" class="nav-link active fs-4 fw-bold" aria-current="page" href="/">Home</a>
+                                </li>
+                                <li class="nav-item">
+                                    <a id="faq-link" class="nav-link disabled fs-4 fw-bold" aria-disabled="true" href="/">FAQs</a>
+                                </li>
+                                <li class="nav-item">
+                                    <a id="about-link" class="nav-link disabled fs-4 fw-bold" aria-disabled="true" href="/">About</a>
+                                </li>
+                                <li class="nav-item">
+                                    <a id="crud-link" class="nav-link fs-4 fw-bold text-center" href="/crud">
+                                        Edit
+                                        <br />
+                                        Bulbs
+                                    </a>
+                                </li>
+                                <li class="nav-item">
+                                    <a
+                                        id="find-bulbs-button"
+                                        class="btn {% if ips|length < 1 %} btn-outline-primary {% else %} btn-primary {% endif %} fw-bolder text-center"
+                                        role="button"
+                                        title="Find Bulbs"
+                                        href="/discover"
+                                    >
+                                        Find
+                                        <br />
+                                        Bulbs
+                                    </a>
+                                </li>
+                            </div>
+                        </ul>
                     </div>
-                    <div class="container-fluid">
-                        {% if numBulbs > 0 %}
-                        <form class="d-flex justify-content-end" id="audio-sync-form" action="javascript:audioSync(this)" method="post">
-                            <select class="form-select me-2" name="audio_device" id="audio_device_selector" title="Pick an audio device from the list">
-                                <option selected>Select an audio device</option>
-                                {% for device in audioDevices %}
-                                <option class="generic-button" value="{{ device.num }}">{{ device.name }}</option>
-                                {% endfor %}
-                            </select>
-                            <button type="submit" class="btn btn-sm btn-outline-danger" id="activate-music-sync-button">Activate music sync</button>
-                        </form>
-                        {% endif %}
-                    </div>
+                </div>
+                <div class="container-fluid">
+                    {% if numBulbs > 0 %}
+                    <form class="d-flex justify-content-end" id="audio-sync-form" action="javascript:audioSync(this)" method="post">
+                        <select class="form-select me-2" name="audio_device" id="audio_device_selector" title="Pick an audio device from the list">
+                            <option selected>Select an audio device</option>
+                            {% for device in audioDevices %}
+                            <option class="generic-button" value="{{ device.num }}">{{ device.name }}</option>
+                            {% endfor %}
+                        </select>
+                        <button type="submit" class="btn btn-sm btn-outline-danger" id="activate-music-sync-button">Activate music sync</button>
+                    </form>
+                    {% endif %}
+                </div>
+                <div class="d-flex justify-content-end">
+                    <i class="fa-solid fa-circle-half-stroke fs-1" onclick="toggleTheme()"></i>
                 </div>
             </div>
         </nav>
+
+        <div id="error-wrapper" class="modal" tabindex="-1">
+            <div class="modal-dialog">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title">Error</h5>
+                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body">
+                        <p>
+                            {% if error %}{{ errorMessage }}
+                            <script>
+                                var errorModal = new bootstrap.Modal(document.getElementById('error-wrapper'));
+                                errorModal.show();
+                            </script>
+                            {% endif %}
+                        </p>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Close</button>
+                    </div>
+                </div>
+            </div>
+        </div>
 
         {% block content %}{% endblock %}
     </body>
